fix: catch error messages
